pyproject.toml
[tool.poetry]
name = "restauranttablereservation"
version = "0.1.0"
description = ""
authors = ["SiMiZZZ"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pillow = "^10.2.0"
environs = "^11.0.0"
alembic = "^1.13.1"
fastapi-restful = "^0.5.0"
asyncpg = "^0.29.0"
typing-inspect = "^0.9.0"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
pydantic = {extras = ["email"], version = "^2.6.4"}
gunicorn = "^21.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
